Rename geometry to primitive_geometry
