+++ ultimaille/primitive_geometry.h
